Start fresh with Prey 2006 game porting

+++ app/src/main/jni/d3es-multithread-master/neo/game/Prey/game_slots.h
@@ -34,7 +34,7 @@
 	}fruit_t;
 
 	typedef struct victory_s {
-		long		f1, f2, f3;
+		int			f1, f2, f3;
 		int			payoff;
 	}victory_t;
